+++ backend/sh/set_iot_leaf_config.sh
@@ -58,35 +58,26 @@ d_echo "w = ${w}"
 
 uuid_gen
 
-p=etc
+p=factory
 read_in_partition
 
-# copy identity config
-d_echo e2cp ${c} /tmp/${uuid}/${p}.img:/upper/aziot/config.toml
-e2mkdir /tmp/${uuid}/${p}.img:/upper/aziot
-e2cp ${c} /tmp/${uuid}/${p}.img:/upper/aziot/config.toml
-
-config_hostname ${c}
-write_back_partition
+copy_identity_config
 
 # create/append to ics_dm_first_boot.sh in factory partition
 # activate identity config on first boot
-p=factory
-read_in_partition
 # for the following cp redirect stderr -> stdout, since it is possible that this file doesnt exist
 e2cp /tmp/${uuid}/${p}.img:/ics_dm_first_boot.sh /tmp/${uuid}/icsd_dm_first_boot.sh 2>&1
 echo "aziotctl config apply" >>  /tmp/${uuid}/ics_dm_first_boot.sh
 e2cp /tmp/${uuid}/ics_dm_first_boot.sh /tmp/${uuid}/${p}.img:/ics_dm_first_boot.sh
+
 write_back_partition
 
 # copy root ca cert
-# @todo refine how we use cert parition
-p=data
+p=cert
 read_in_partition
-d_echo e2cp ${r} /tmp/${uuid}/${p}.img:/local/share/ca-certificates/$(basename ${r}).crt
-e2mkdir /tmp/${uuid}/${p}.img:/local/share/ca-certificates
-e2cp ${r} /tmp/${uuid}/${p}.img:/local/share/ca-certificates/$(basename ${r}).crt
+
+d_echo e2cp ${r} /tmp/${uuid}/${p}.img:/ca/$(basename ${r}).crt
+e2mkdir /tmp/${uuid}/${p}.img:/ca
+e2cp ${r} /tmp/${uuid}/${p}.img:/ca/$(basename ${r}).crt
+
 write_back_partition
-# Just a remark: copying the root ca cert isn't sufficient. The device has
-# to call update-ca-certificates on first boot ... we handle that in
-# ics-dm-first-boot.service