+++ backend/sh/set_wifi_config.sh
@@ -39,22 +39,19 @@ d_echo "w = ${w}"
 
 uuid_gen
 
-p=etc
+p=factory
 read_in_partition
 
 # copy wpa_supplicant conf
-d_echo "e2cp ${i} /tmp/${uuid}/${p}.img:/upper/wpa_supplicant/wpa_supplicant-wlan0.conf"
-e2mkdir /tmp/${uuid}/${p}.img:/upper/wpa_supplicant
-e2cp -P 644 ${i} /tmp/${uuid}/${p}.img:/upper/wpa_supplicant/wpa_supplicant-wlan0.conf
-
-write_back_partition
+d_echo "e2cp ${i} /tmp/${uuid}/${p}.img:/etc/wpa_supplicant/wpa_supplicant-wlan0.conf"
+e2mkdir /tmp/${uuid}/${p}.img:/etc/wpa_supplicant
+e2cp -P 644 ${i} /tmp/${uuid}/${p}.img:/etc/wpa_supplicant/wpa_supplicant-wlan0.conf
 
 # enable wpa_supplicant
 # create/append to ics_dm_first_boot.sh in factory partition
-p=factory
-read_in_partition
 # for the following cp redirect stderr -> stdout, since it is possible that this file doesnt exist
 e2cp /tmp/${uuid}/${p}.img:/ics_dm_first_boot.sh /tmp/${uuid}/icsd_dm_first_boot.sh 2>&1
 echo "systemctl enable wpa_supplicant@wlan0.service && systemctl start wpa_supplicant@wlan0.service" >> /tmp/${uuid}/ics_dm_first_boot.sh
 e2cp /tmp/${uuid}/ics_dm_first_boot.sh /tmp/${uuid}/${p}.img:/ics_dm_first_boot.sh
+
 write_back_partition